refactor: consolidate inline frontend types into shared api-interfaces lib (#154)
- Move ~30 inline frontend types into @chansey/api-interfaces shared lib
- Fix Order interface: make fee/commission required, baseCoin/quoteCoin nullable
- Fix Coin/Exchange: make createdAt optional (select: false), rename logo to image
- Expand UpdateExchangeDto to match all backend CreateExchangeDto fields
- Add PortfolioType enum, fix portfolio timestamp types to optional Date
- Move OtpResponse to otp.interface.ts alongside related OTP types
- Remove ghost field preferred_username from IUserProfileUpdate
- Add id field to UserAsset to match backend AssetDetailsDto
- Fix OrderType underscore display in transactions table and filter dropdown
- Fix resize event listener memory leak in exchange-balance component

Closes #154
