+++ apps/chansey/src/app/pages/admin/categories/categories.component.ts
@@ -1,5 +1,5 @@
 import { CommonModule } from '@angular/common';
-import { Component, ViewChild, ElementRef, computed, effect, inject, signal } from '@angular/core';
+import { Component, ElementRef, ViewChild, computed, effect, inject, signal } from '@angular/core';
 import { FormBuilder, FormGroup, FormsModule, ReactiveFormsModule, Validators } from '@angular/forms';
 
 import { ConfirmationService, MessageService } from 'primeng/api';
@@ -15,7 +15,9 @@ import { InputTextModule } from 'primeng/inputtext';
 import { Table, TableModule } from 'primeng/table';
 import { ToastModule } from 'primeng/toast';
 
-import { Category, CategoriesService } from './categories.service';
+import { Category } from '@chansey/api-interfaces';
+
+import { CategoriesService } from './categories.service';
 
 @Component({
   selector: 'app-categories',
