+++ apps/chansey/src/app/pages/transactions/transactions.service.ts
@@ -2,75 +2,8 @@ import { Injectable, Signal } from '@angular/core';
 
 import { injectQuery } from '@tanstack/angular-query-experimental';
 
-import { queryKeys, useAuthQuery, authenticatedFetch, STANDARD_POLICY, FREQUENT_POLICY } from '@chansey/shared';
-
-// Order types from the backend
-export enum OrderType {
-  LIMIT = 'LIMIT',
-  LIMIT_MAKER = 'LIMIT_MAKER',
-  MARKET = 'MARKET',
-  STOP = 'STOP',
-  STOP_MARKET = 'STOP_MARKET',
-  STOP_LOSS_LIMIT = 'STOP_LOSS_LIMIT',
-  TAKE_PROFIT_LIMIT = 'TAKE_PROFIT_LIMIT',
-  TAKE_PROFIT_MARKET = 'TAKE_PROFIT_MARKET',
-  TRAILING_STOP_MARKET = 'TRAILING_STOP_MARKET'
-}
-
-export enum OrderSide {
-  BUY = 'BUY',
-  SELL = 'SELL'
-}
-
-export enum OrderStatus {
-  CANCELED = 'CANCELED',
-  EXPIRED = 'EXPIRED',
-  FILLED = 'FILLED',
-  NEW = 'NEW',
-  PARTIALLY_FILLED = 'PARTIALLY_FILLED',
-  PENDING_CANCEL = 'PENDING_CANCEL',
-  REJECTED = 'REJECTED'
-}
-
-export interface Transaction {
-  id: string;
-  symbol: string;
-  orderId: string;
-  clientOrderId: string;
-  transactTime: Date;
-  quantity: number;
-  price: number;
-  executedQuantity: number;
-  status: OrderStatus;
-  side: OrderSide;
-  type: OrderType;
-  cost?: number;
-  fee: number;
-  commission: number;
-  feeCurrency?: string;
-  baseCoin: {
-    id: string;
-    name: string;
-    symbol: string;
-    slug: string;
-    image: string;
-  };
-  quoteCoin: {
-    id: string;
-    name: string;
-    symbol: string;
-    slug: string;
-    image: string;
-  };
-  exchange?: {
-    id: string;
-    name: string;
-    slug: string;
-    image?: string;
-  };
-  createdAt: Date;
-  updatedAt: Date;
-}
+import { Order } from '@chansey/api-interfaces';
+import { authenticatedFetch, FREQUENT_POLICY, queryKeys, STANDARD_POLICY, useAuthQuery } from '@chansey/shared';
 
 /**
  * Service for transactions/orders data via TanStack Query
@@ -87,7 +20,7 @@ export class TransactionsService {
    * Query all transactions
    */
   useTransactions() {
-    return useAuthQuery<Transaction[]>(queryKeys.transactions.lists(), this.apiUrl, {
+    return useAuthQuery<Order[]>(queryKeys.transactions.lists(), this.apiUrl, {
       cachePolicy: STANDARD_POLICY
     });
   }
@@ -102,7 +35,7 @@ export class TransactionsService {
       const id = transactionId();
       return {
         queryKey: queryKeys.transactions.detail(id || ''),
-        queryFn: () => authenticatedFetch<Transaction>(`${this.apiUrl}/${id}`),
+        queryFn: () => authenticatedFetch<Order>(`${this.apiUrl}/${id}`),
         ...STANDARD_POLICY,
         enabled: !!id
       };
@@ -115,7 +48,7 @@ export class TransactionsService {
    * Uses FREQUENT policy since open orders may change often
    */
   useOpenTransactions() {
-    return useAuthQuery<Transaction[]>(queryKeys.transactions.open(), `${this.apiUrl}/open`, {
+    return useAuthQuery<Order[]>(queryKeys.transactions.open(), `${this.apiUrl}/open`, {
       cachePolicy: FREQUENT_POLICY
     });
   }
