+++ apps/chansey/src/app/shared/components/recent-transactions/recent-transactions.component.ts
@@ -9,7 +9,9 @@ import { SkeletonModule } from 'primeng/skeleton';
 import { TableModule } from 'primeng/table';
 import { TagModule } from 'primeng/tag';
 
-import { OrderSide, OrderStatus, TransactionsService } from '@chansey-web/app/pages/transactions/transactions.service';
+import { OrderSide, OrderStatus } from '@chansey/api-interfaces';
+
+import { TransactionsService } from '@chansey-web/app/pages/transactions/transactions.service';
 
 @Component({
   selector: 'app-recent-transactions',
