+++ libs/api-interfaces/src/index.ts
@@ -4,11 +4,14 @@ export * from './lib/algorithm/algorithm.interface';
 export * from './lib/api-interfaces';
 export * from './lib/audit/audit-entry.interface';
 export * from './lib/auth';
+export * from './lib/balance';
+export * from './lib/category';
 export * from './lib/coin/coin.interface';
 export * from './lib/coin/ticker-pair-status.enum';
 export * from './lib/coin/ticker-pair.interface';
 export * from './lib/exchange/exchange.interface';
 export * from './lib/order';
+export * from './lib/portfolio';
 export * from './lib/risk';
 export * from './lib/strategy/strategy-config.interface';
 export * from './lib/user';
