+++ libs/api-interfaces/src/lib/auth/index.ts
@@ -1,11 +1,12 @@
+export * from './forgot-password-response.interface';
 export * from './forgot-password.interface';
 export * from './login-response.interface';
 export * from './login.interface';
 export * from './logout-response.interface';
-export * from './register.interface';
-export * from './register-response.interface';
-export * from './forgot-password-response.interface';
-export * from './resend-email-response.interface';
-export * from './reset-password.interface';
-export * from './reset-password-response.interface';
 export * from './otp.interface';
+export * from './register-response.interface';
+export * from './register.interface';
+export * from './resend-email-response.interface';
+export * from './reset-password-response.interface';
+export * from './reset-password.interface';
+export * from './verify-email.interface';
