+++ libs/api-interfaces/src/lib/auth/verify-email.interface.ts
@@ -1,0 +1,7 @@
+export interface VerifyEmailRequest {
+  token: string;
+}
+
+export interface VerifyEmailResponse {
+  message: string;
+}
