+++ libs/api-interfaces/src/lib/balance/index.ts
@@ -1,0 +1,1 @@
+export * from './balance.interface';
