+++ libs/api-interfaces/src/lib/category/category.interface.ts
@@ -1,0 +1,18 @@
+export interface Category {
+  id: string;
+  name: string;
+  slug: string;
+  createdAt: Date;
+  updatedAt: Date;
+}
+
+export interface CreateCategoryDto {
+  name: string;
+  slug: string;
+}
+
+export interface UpdateCategoryDto {
+  id: string;
+  name?: string;
+  slug?: string;
+}
