+++ libs/api-interfaces/src/lib/category/index.ts
@@ -1,0 +1,1 @@
+export * from './category.interface';
