+++ libs/api-interfaces/src/lib/portfolio/index.ts
@@ -1,0 +1,2 @@
+export * from './portfolio-type.enum';
+export * from './portfolio.interface';
