+++ libs/api-interfaces/src/lib/portfolio/portfolio-type.enum.ts
@@ -1,0 +1,4 @@
+export enum PortfolioType {
+  MANUAL = 'MANUAL',
+  AUTOMATIC = 'AUTOMATIC'
+}
