+++ libs/api-interfaces/src/lib/portfolio/portfolio.interface.ts
@@ -1,0 +1,16 @@
+import { PortfolioType } from './portfolio-type.enum';
+
+import { Coin } from '../coin/coin.interface';
+
+export interface CreatePortfolioDto {
+  coinId: string;
+  type: PortfolioType;
+}
+
+export interface PortfolioItem {
+  id: string;
+  coin: Coin;
+  type: PortfolioType;
+  createdAt?: Date;
+  updatedAt?: Date;
+}
